tools: create mdsdir frag in cephfs-data-scan init
MDS used to ignore missing dirfrag objects.  Now that
they're detected and flagged, we need to explicitly
inject them during disaster recovery in order to have
a working filesystem.

(Specifically, this is needed because in CDir::_go_bad
we used to finish_waiting(WAIT_COMPLETE, 0), but now
we finish_waiting(WAIT_COMPLETE, -EIO))

Signed-off-by: John Spray <john.spray@redhat.com>