auth: also catch end_of_buffer while parsing text input
[On FreeBSD]
in KeyRing::decode() text parsing in only executed if decode throws.
This results in an ABORT signal in bin/ceph when running vstart.sh

Only buffer::error is caught, but reading config text sometimes
lets the buffer_length run into absurd numbers.
Getting out of bounds access in the buffer, generating a trap.

So we need to catch that as well.

However adding buffer::end_of_buffer as exception in another case
does not seem to work, because it is not matched????
So catching all exceptions, and continue to perhaps trap in  decode_plaintext()

fixes: https://tracker.ceph.com/issues/41717
Signed-off-by: Willem Jan Withagen <wjw@digiware.nl>